+++ tests/com/goide/psi/legacy/GoLegacyResolveTypesTest.java
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package com.goide.psi;
-
-import com.goide.GoLegacyResolveTestBase;
+package com.goide.psi.legacy;
 
 public class GoLegacyResolveTypesTest extends GoLegacyResolveTestBase {
   @Override
